requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
Pillow>=9.0.0
numpy>=1.21.0
requests>=2.28.0
opencv-python>=4.7.0
accelerate>=0.20.0
sentencepiece>=0.1.99
timm>=0.9.0 